fix non low code scanning warnings and tab loading
